% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATNeu_example}
\alias{ATNeu_example}
\title{Sample Data}
\format{
A list containing
\itemize{
\item \code{forcing} as an \code{\link[xts:xts]{xts::xts()}} object, and
\item \code{soil} as a named double vector.
}
}
\usage{
ATNeu_example
}
\description{
Sample Data
}
\keyword{datasets}
