% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bourne}
\alias{Bourne}
\title{Bourne SNOTEL station sample data}
\format{
A list containing
\itemize{
\item \code{forcing} as an \code{\link[xts:xts]{xts::xts()}} object, and
\item \code{max_depth_sm} numeric vector length 1
\item \code{soil} as a named double vector.
\item \code{md} a metadata \code{\link[=data.frame]{data.frame()}} with 1 row and 28 columns
}
}
\source{
\url{https://wcc.sc.egov.usda.gov/nwcc/site?sitenum=361}
}
\usage{
Bourne
}
\description{
Bourne SNOTEL station sample data
}
\keyword{datasets}
