% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SA_cru}
\alias{SA_cru}
\title{South American sample gridded data}
\format{
A list containing
\itemize{
\item \code{tc} a \link[raster:Raster-classes]{RasterBrick} of monthly temperature data (mean deg C)
\item \code{pn} a \link[raster:Raster-classes]{RasterBrick} of monthly precipitation data (mean mm/d or mm/month?)
\item \code{sw_in} a \link[raster:Raster-classes]{RasterBrick} of monthly precipitation data (mean W m^-2)
\item \code{elev} a \link[raster:Raster-classes]{RasterLayer} representing a DEM over South America (presumably in average meters above sealevel)
\item \code{soil} a \link[raster:Raster-classes]{RasterBrick} representing the soil fractions
\itemize{
\item ⚠️ sometimes adding up to less, sometimes to more than 100\% ⚠️
}
}
}
\source{
\url{https://crudata.uea.ac.uk/cru/data/hrg/}
}
\usage{
SA_cru
}
\description{
Used to construct a \code{\link[=splash.grid]{splash.grid()}} call. The CRU in the name probably stands for the
\href{https://research-portal.uea.ac.uk/en/organisations/climatic-research-unit}{Climatic Research Unit} at
the University of East Anglia. The soil data's
}
\keyword{datasets}
