% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsSWC.grid.R
\name{unSWC.grid}
\alias{unSWC.grid}
\title{unsSWC.grid}
\usage{
unSWC.grid(soil_data, uns_depth, wn, d_name, outdir = getwd())
}
\arguments{
\item{soil_data, }{Raster* object with the layers organized as sand(perc),clay(perc),organic matter(perc),coarse-fragments-fraction(perc), bulk density(g cm-3)}

\item{uns_depth, }{maximum depth to calculate the accumulated water content in (m)}

\item{wn, }{Raster* object, total water content of the profile from surface to bedrock (m)}

\item{outdir, }{directory where to save the files}
}
\description{
Estimates water content in the unsaturated part of the profile using an analytical integral of the eqn. described by Brooks and Corey 1964 and calculates depth to water table if there is any (*experimental)
}
\keyword{splashTools}
