% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsSWC.R
\name{unSWC}
\alias{unSWC}
\title{unsSWC}
\usage{
unSWC(soil_data, uns_depth, wn, units = "mm")
}
\arguments{
\item{soil_data, }{vector as retrieved from splashTools::getSoilNRCS() or splashTools::getSoilSite()}

\item{uns_depth, }{maximum depth to calculate the accumulated water content in (m)}

\item{wn, }{xts object, total water content of the profile from surface to bedrock (m)}

\item{ouputdir, }{directory where to save the files}
}
\description{
Estimates water content using an analytca integral of the profile described by Brooks and Corey 1964 and calculates depth to water table if there is any
}
\examples{
unSWC()
}
\keyword{splashTools}
