% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_splash_grid.R
\name{splash.grid}
\alias{splash.grid}
\title{splash.grid}
\usage{
splash.grid(sw_in, tc, pn, elev, soil, outdir = getwd(),
  sim.control = list(par = TRUE, ncores = 7, output.mode = "monthly",
  inmem = FALSE), ...)
}
\arguments{
\item{sw_in, }{lon}

\item{tc, }{lon}

\item{pn, }{lon}

\item{elev, }{lon}
}
\value{
a matrix xts type
}
\description{
Apply splash algorithm
}
\examples{
splash.grid()
}
\keyword{splash}
