% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash.point.R
\name{splash.point}
\alias{splash.point}
\title{Simple process-led algorithms for simulating habitats (SPLASH v.2.0)}
\usage{
splash.point(sw_in, tc, pn, lat, elev, slop = 0, asp, soil_data,
  Au = 0, resolution = 250)
}
\arguments{
\item{sw_in}{Incoming shortwave solar radiation (W m-2), timeseries object of monthly or daily averages.}

\item{tc}{Air temperature (戼㸰C), same timestep as sw_in}

\item{pn}{Precipitation (mm), same timestep as sw_in}

\item{elev}{Elevation (m.a.s.l)}

\item{slop}{Terrain feature: slope inclination (戼㸰)}

\item{asp}{Terrain feature: slope orientation (戼㸰), standard clockwise from 0.0戼㸰 North}

\item{soil_data}{Soil data organized as a vector in the way: c(sand(%),clay(%),organic matter(%),coarse-fragments-fraction(%), bulk density(g cm-3))}
}
\value{
a time series matrix including:
\itemize{
        \item \eqn{W_n}: Soil water content within the first 2m of depth (mm).
        \item \eqn{ro}: Runoff (mm d-1).
        \item \eqn{pet}: Potential evapotranspiration (mm d-1).
        \item \eqn{aet}: Actual evapotranspiration (mm d-1).
        \item \eqn{snow}: Snow water equivalent (mm).
        \item \eqn{cond}: Condensation (mm d-1).
        \item \eqn{bflow}: Lateral flow (mm d-1).
        \item \eqn{netr}: Daytime net radiation (MJ d-1).
}
}
\description{
R/C++ implementation of the SPLASH v.2.0 algorithm (Davis et al., 2017; Sandoval et al., in prep.).
}
\examples{
splash.point(sw_in=200, tc=15, pn=10, lat=44,elev=1800,slop=10,asp=270,soil_data=c(sand=44,clay=2,OM=6,fgravel=12))
}
\keyword{evapotranspiration,}
\keyword{moisture}
\keyword{soil}
\keyword{splash,}
