% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsSWC.grid.R
\name{unSWC.grid}
\alias{unSWC.grid}
\title{unsSWC.grid}
\usage{
unSWC.grid(soil_data, uns_depth, wn)
}
\arguments{
\item{soil_data, }{Raster* object with the layers organized as sand(%),clay(%),organic matter(%),coarse-fragments-fraction(%), bulk density(g cm-3)}

\item{uns_depth, }{maximum depth to calculate the accumulated water content in (m)}

\item{wn, }{Raster* object, total water content of the profile from surface to bedrock (m)}

\item{ouputdir, }{directory where to save the files}
}
\description{
Estimates water content using an analytca integral of the profile described by Brooks and Corey 1964 and calculates depth to water table if there is any
}
\examples{
unSWC()
}
\keyword{splashTools}
