% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upslope_area.R
\name{upslope_area}
\alias{upslope_area}
\title{upslope_area}
\usage{
upslope_area(dem)
}
\arguments{
\item{dem}{}
}
\value{
a RasterStack with the contributing area per pixel, number of cells draining in, and number of cells draining out
}
\description{
Computes the upslope area in m2 using topmodel for small areas or Taudem for large, corrections acording to latitude done by raster::area
}
\examples{
splash.grid()
}
\keyword{area}
\keyword{contributing}
