% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upslope_area.R
\name{upslope_area}
\alias{upslope_area}
\title{upslope_area}
\usage{
upslope_area(dem)
}
\arguments{
\item{sw_in, }{lon}

\item{tc, }{lon}

\item{pn, }{lon}

\item{elev, }{lon}
}
\value{
a matrix xts type
}
\description{
Computes the upslope area in m2 using topmodel for small areas or Taudem for large
}
\examples{
splash.grid()
}
\keyword{splash}
